Feat : 버티컬 컬러를 이용해서 지층 색 그라데이션 적용
- VertexColorShader를 이용하여 VoxelTerrainMash의 색상 적용할 수 있게 추가
- VoxelTerrain 코드에 그라데이션 할 수 있게 수정

+++ Assets/02.Scripts/10.Voxel/VoxelSurfaceGenerator.cs
@@ -56,7 +56,7 @@ public class VoxelSurfaceGenerator : MonoBehaviour
             }
         }
     }
-    public float GetSurfaceHeight(int x, int z)
+    public float GetSurfaceHeight(float x, float z)
     {
         return baseTerrainHeight + Mathf.PerlinNoise(x * terrainScale, z * terrainScale) * mountainHeight;
     }
